js packages introduced and jest configured
