+++ .eslintrc.json
@@ -2,7 +2,8 @@
     "env": {
         "browser": true,
         "commonjs": true,
-        "es2021": true
+        "es2021": true,
+        "jest": true
     },
     "extends": [
         "airbnb-base"
@@ -12,6 +13,7 @@
     },
     "rules": {
         "no-console": ["error", { "allow": ["log", "warn", "error"] }],
-        "indent": ["error", 4]
+        "indent": ["error", 4],
+        "import/extensions": ["always"]
     }
 }
